Search with removing code repeatation and enable the SS, Video on failure 🥸
